Remove gRPC code from the codebase
[skip ci]
There are no reasonable use cases for gRPC,
since raw sockets can do everything that gRPC
can, but has much better CPU consumption.
Thus, all code related to gRPC can be removed.
