+++ doc/compiler/jitaas/README.md
@@ -1,5 +1,5 @@
 <!--
-Copyright (c) 2018, 2018 IBM Corp. and others
+Copyright (c) 2000, 2019 IBM Corp. and others
 
 This program and the accompanying materials are made available under
 the terms of the Eclipse Public License 2.0 which accompanies this
@@ -38,4 +38,3 @@ SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-excepti
 - [CHTable](CHTable.md)
 - [IProfiler](IProfiler.md)
 - [Client Sessions](ClientSession.md)
-- [gRPC (deprecated)](gRPC.md)
