+++ doc/compiler/jitaas/Sockets.md
@@ -1,5 +1,5 @@
 <!--
-Copyright (c) 2018, 2018 IBM Corp. and others
+Copyright (c) 2000, 2019 IBM Corp. and others
 
 This program and the accompanying materials are made available under
 the terms of the Eclipse Public License 2.0 which accompanies this
@@ -22,7 +22,7 @@ SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-excepti
 
 # Sockets
 
-Client server communication via TCP sockets is the default communication backend. The implementation mostly resides within the `rpc/raw` directory. There is a base class `J9Stream`, which is specialized for both the client and server in `J9ClientStream` and `J9ServerStream`.
+Client server communication via TCP sockets is the default communication backend. The implementation mostly resides within the `rpc` directory. There is a base class `J9Stream`, which is specialized for both the client and server in `J9ClientStream` and `J9ServerStream`.
 
 Encryption via TLS (OpenSSL) is optionally supported. See [Usage](Usage.md) for encryption setup instructions.
