+++ runtime/compiler/rpc/J9Client.cpp
@@ -20,8 +20,6 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#ifdef JITAAS_USE_RAW_SOCKETS
-
 #include "J9Client.hpp"
 #include "control/Options.hpp"
 #include "env/VerboseLog.hpp"
@@ -273,5 +271,3 @@ J9ClientStream::shutdown()
    }
 
 };
-
-#endif
