+++ runtime/compiler/rpc/J9Client.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2018 IBM Corp. and others
+ * Copyright (c) 2018, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -20,10 +20,75 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#ifdef JITAAS_USE_GRPC
-#include "grpc/J9Client.hpp"
-#endif
+#ifndef J9_CLIENT_H
+#define J9_CLIENT_H
 
-#ifdef JITAAS_USE_RAW_SOCKETS
-#include "raw/J9Client.hpp"
-#endif
+#include <google/protobuf/io/zero_copy_stream_impl.h>
+#include <openssl/ssl.h>
+#include "compile/CompilationTypes.hpp"
+#include "rpc/StreamTypes.hpp"
+#include "rpc/ProtobufTypeConvert.hpp"
+#include "j9.h"
+#include "ilgen/J9IlGeneratorMethodDetails.hpp"
+#include "rpc/J9Stream.hpp"
+
+class SSLOutputStream;
+class SSLInputStream;
+
+namespace JITaaS
+{
+class J9ClientStream : J9Stream
+   {
+public:
+   static void static_init(TR::PersistentInfo *info);
+
+   J9ClientStream(TR::PersistentInfo *info);
+   virtual ~J9ClientStream()
+      {
+      _numConnectionsClosed++;
+      }
+
+   template <typename... T>
+   void buildCompileRequest(T... args)
+      {
+      write(args...);
+      }
+
+   Status waitForFinish();
+
+   template <typename ...T>
+   void write(T... args)
+      {
+      _cMsg.set_status(true);
+      setArgs<T...>(_cMsg.mutable_data(), args...);
+      writeBlocking(_cMsg);
+      }
+
+   void writeError();
+
+   J9ServerMessageType read()
+      {
+      readBlocking(_sMsg);
+      return _sMsg.type();
+      }
+
+   template <typename ...T>
+   std::tuple<T...> getRecvData()
+      {
+      return getArgs<T...>(_sMsg.mutable_data());
+      }
+
+   void shutdown();
+
+   static int _numConnectionsOpened;
+   static int _numConnectionsClosed;
+
+private:
+   uint32_t _timeout;
+
+   static SSL_CTX *_sslCtx;
+   };
+
+}
+
+#endif // J9_CLIENT_H
