+++ runtime/compiler/rpc/J9Server.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2018 IBM Corp. and others
+ * Copyright (c) 2000, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -20,10 +20,89 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#ifdef JITAAS_USE_GRPC
-#include "grpc/J9Server.hpp"
-#endif
+#ifndef J9_SERVER_H
+#define J9_SERVER_H
 
-#ifdef JITAAS_USE_RAW_SOCKETS
-#include "raw/J9Server.hpp"
-#endif
+#include <google/protobuf/io/zero_copy_stream_impl.h>
+#include <openssl/ssl.h>
+#include "rpc/ProtobufTypeConvert.hpp"
+#include "rpc/J9Stream.hpp"
+#include "env/CHTable.hpp"
+
+class SSLOutputStream;
+class SSLInputStream;
+
+namespace JITaaS
+{
+class J9ServerStream : J9Stream
+   {
+public:
+   J9ServerStream(int connfd, BIO *ssl, uint32_t timeout);
+   virtual ~J9ServerStream() 
+      {
+      _numConnectionsClosed++;
+      }
+
+   template <typename ...T>
+   void write(J9ServerMessageType type, T... args)
+      {
+      setArgs<T...>(_sMsg.mutable_data(), args...);
+      _sMsg.set_type(type);
+      writeBlocking(_sMsg);
+      }
+   template <typename ...T>
+   std::tuple<T...> read()
+      {
+      readBlocking(_cMsg);
+      if (!_cMsg.status())
+         throw StreamCancel();
+      return getArgs<T...>(_cMsg.mutable_data());
+      }
+
+   void cancel();
+   void finishCompilation(uint32_t statusCode, std::string codeCache = "", std::string dataCache = "", CHTableCommitData chTableData = {}, std::vector<TR_OpaqueClassBlock*> classesThatShouldNotBeNewlyExtended = {}, std::string logFileStr = "", std::string symbolToIdStr = "");
+   void setClientId(uint64_t clientId)
+      {
+      _clientId = clientId;
+      }
+   uint64_t getClientId() const
+      {
+      return _clientId;
+      }
+
+   static int _numConnectionsOpened;
+   static int _numConnectionsClosed;
+
+private:
+   void finish();
+
+   uint32_t _msTimeout;
+   uint64_t _clientId;
+   };
+
+//
+// Inherited class is starting point for the received compilation request
+//
+class J9BaseCompileDispatcher
+   {
+public:
+   virtual void compile(J9ServerStream *stream) = 0;
+   };
+
+class J9CompileServer
+   {
+public:
+   ~J9CompileServer()
+      {
+      }
+
+   void buildAndServe(J9BaseCompileDispatcher *compiler, TR::PersistentInfo *info);
+
+private:
+   void serve(J9BaseCompileDispatcher *compiler, uint32_t timeout);
+   bool waitOnQueue();
+   };
+
+}
+
+#endif // J9_SERVER_H
