+++ runtime/compiler/rpc/ProtobufTypeConvert.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2018 IBM Corp. and others
+ * Copyright (c) 2000, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -35,7 +35,7 @@ namespace JITaaS
    template <typename... Args>
    void setArgs(AnyData *message, Args... args);
 
-   // Sending void over gRPC is awkward, so we just wrap a bool...
+   // Sending void over protobuf is awkward, so we just wrap a bool...
    struct Void {
       Void() {}
       Void(bool) {}
